Extract common oods code to new file.
