+++ src/core/mod.rs
@@ -5,4 +5,5 @@ pub mod constraints;
 pub mod fft;
 pub mod fields;
 pub mod fri;
+pub mod oods;
 pub mod poly;
